Added ability to push toasts [bugged]
Unfortunately I might have to scrap this idea. I can't figure out how to trigger a `RedrawRequested` event after an operation is processed. This means that it is possible that we successfully add a toast, but it won't show up until we send another event (e.g. we move the mouse a little). I believe this might be a limitation of the system, and I will have to try another way.