Cleaned up API
Added the builder pattern for toasts so users can optionally add the features they want whilst still keeping a clean interface. Also made naming more specific: I.e. ToastLevel and ToastId instead of just Level and Id, which could have caused naming conflicts. Renamed ToastManager to ToastContainer, which is a more useful in helping the user understand it's user case (it's a container, so it holds contents inside of it)